refactored classNameResolver
