--- src/tw-styled/tools.ts
@@ -1,1 +1,0 @@
-export * from "./style-to-class-resolver";
