--- src/tw-styled/types/index.ts
@@ -1,1 +1,0 @@
-export * from "../style-to-class-resolver/types";
