+++ src/ui/classNameResolver/buckets/animation.ts
@@ -1,4 +1,4 @@
-import { Keyof, SingleBucket } from "../types";
+import { Keyof, SingleBucket } from "../../types";
 
 enum AnimationBucket {
   spin = "animate-spin",
