+++ src/ui/classNameResolver/buckets/color.ts
@@ -1,4 +1,4 @@
-import { Keyof, MultiBucket } from "../types";
+import { Keyof, MultiBucket } from "../../types";
 
 export enum Color {
   background = "background",
