+++ src/ui/components/box/Box.tsx
@@ -1,12 +1,12 @@
 "use client";
 import React from "react";
-import { useStyleToClass } from "@/tw-styled/style-to-class-resolver";
+import { useClassNameResolver } from "@/ui";
 import { BoxProps } from "./Box.types";
 
 export const Box = (props: BoxProps) => {
   const { as = "div", style, children } = props;
 
-  const classes = useStyleToClass({
+  const classes = useClassNameResolver({
     wrapper: { ...style }
   });
