+++ src/ui/components/form/Form.types.ts
@@ -1,4 +1,4 @@
-import { BaseStyles } from "@/tw-styled/types";
+import { BaseStyles } from "@/ui/types";
 import { HeadingElements } from "@/ui/types";
 
 export type FormProps = {
