+++ src/ui/components/heading/Heading.types.tsx
@@ -1,4 +1,4 @@
-import { BaseStyles } from "@/tw-styled/types";
+import { BaseStyles } from "@/ui/types";
 import { HeadingElements } from "@/ui/types";
 
 export type HeadingProps = {
