+++ src/ui/components/icon/Icon.types.ts
@@ -1,5 +1,5 @@
 import { IconNames, IconVariants } from "@/shared/hooks";
-import { BaseStyles } from "@/tw-styled/types";
+import { BaseStyles } from "@/ui/types";
 
 export type IconProps = {
   name: IconNames;
