+++ src/ui/components/input/Input.types.ts
@@ -1,4 +1,4 @@
-import { BaseStyles } from "@/tw-styled/types";
+import { BaseStyles } from "@/ui/types";
 import { FieldValues, UseFormRegister } from "react-hook-form";
 
 type InputType = "text" | "password" | "checkbox" | "textarea";
