+++ src/ui/components/layout/Layout.types.ts
@@ -1,4 +1,4 @@
-import { BaseStyles } from "@/tw-styled/types";
+import { BaseStyles } from "@/ui/types";
 import { WrapperElements } from "@/ui/types";
 
 export type LayoutProps = {
