+++ src/ui/components/pulse/Pulse.types.ts
@@ -1,4 +1,4 @@
-import { BaseStyles } from "@/tw-styled/types";
+import { BaseStyles } from "@/ui/types";
 
 export type PulseProps = {
   size?: "sm" | "md" | "lg";
