+++ src/ui/components/spin/Spin.types.ts
@@ -1,4 +1,4 @@
-import { BaseStyles } from "@/tw-styled/types";
+import { BaseStyles } from "@/ui/types";
 
 export type SpinProps = {
   style?: {
